+++ src/main/java/project/linkarchive/backend/url/response/RefactorUserLinkList/LinkResponse.java
@@ -4,7 +4,7 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 
 @Getter
-public class UrlResponse {
+public class LinkResponse {
     private Long urlId;
     private String link;
     private String title;
@@ -13,7 +13,7 @@ public class UrlResponse {
     private Long bookMarkCount;
 
     @QueryProjection
-    public UrlResponse(Long urlId, String link, String title, String description, String thumbnail, Long bookMarkCount) {
+    public LinkResponse(Long urlId, String link, String title, String description, String thumbnail, Long bookMarkCount) {
         this.urlId = urlId;
         this.link = link;
         this.title = title;
